Fix REPL, make REPL use resolution
